# web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Tropicast.Dashboard.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Signs a desktop device in: access and refresh token.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/devices/{id}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Signs one desktop device out and revokes the broadcast credentials it got for itself.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/auth/device/code": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Starts a desktop sign-in: a secret device code to poll with and a short user code to show.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCodeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/device/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Polls a desktop sign-in; returns the device's tokens once the user approved it.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/device/{userCode}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "A pending desktop sign-in, to confirm before approving it.",
        "parameters": [
          {
            "name": "userCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/device/approve": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Signs the device with this code in to the signed-in account.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/auth/device/deny": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refuses the desktop sign-in with this code.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceCodeRequest": {
        "required": [
          "deviceName"
        ],
        "type": "object",
        "properties": {
          "deviceName": {
            "maxLength": 64,
            "type": "string"
          }
        }
      },
      "DeviceCodeResponse": {
        "required": [
          "deviceCode",
          "userCode",
          "verificationUri",
          "verificationUriComplete",
          "expiresIn",
          "interval"
        ],
        "type": "object",
        "properties": {
          "deviceCode": {
            "type": "string",
            "description": "Secret: send it to `/auth/device/token` only. Never show it."
          },
          "userCode": {
            "type": "string",
            "description": "Short code the user types on the web page, e.g. `BCDF-GHJK`."
          },
          "verificationUri": {
            "type": "string",
            "description": "Web page where the user enters the code.",
            "format": "uri"
          },
          "verificationUriComplete": {
            "type": "string",
            "description": "The same page with the code filled in, e.g. for a button or a QR code.",
            "format": "uri"
          },
          "expiresIn": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Seconds until the codes expire.",
            "format": "int64"
          },
          "interval": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Minimum seconds between two polls.",
            "format": "int64"
          }
        },
        "description": "A started desktop sign-in. Show string DeviceCodeResponse.UserCode and Uri DeviceCodeResponse.VerificationUri, then poll."
      },
      "DeviceRequestResponse": {
        "required": [
          "userCode",
          "deviceName",
          "createdAt",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "userCode": {
            "type": "string",
            "description": "The code, as the device shows it."
          },
          "deviceName": {
            "type": "string",
            "description": "Name the device gave itself; it becomes the device's name in the account."
          },
          "createdAt": {
            "type": "string",
            "description": "When the device asked.",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "description": "When the request expires.",
            "format": "date-time"
          }
        },
        "description": "A pending desktop sign-in, shown to the user before they approve it."
      },
...
      "DeviceTokenRequest": {
        "required": [
          "deviceCode"
        ],
        "type": "object",
        "properties": {
          "deviceCode": {
            "maxLength": 256,
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "TokenRequest": {
        "required": [
          "email",
          "password",
          "deviceName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 256,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "deviceName": {
            "maxLength": 64,
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "required": [
          "tokenType",
          "accessToken",
          "expiresIn",
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "tokenType": {
            "type": "string",
            "description": "Always `Bearer`."
          },
          "accessToken": {
            "type": "string",
            "description": "Send as `Authorization: Bearer …`."
          },
          "expiresIn": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Access token lifetime in seconds.",
            "format": "int64"
          },
          "refreshToken": {
            "type": "string",
            "description": "One-time: each refresh returns a new one. Revoke it to sign the device out."
          }
        },
        "description": "Tokens for the desktop app: string TokenResponse.AccessToken lasts 15 minutes; refresh to get a new pair."
      },
...
      "UserCodeRequest": {
        "required": [
          "userCode"
        ],
        "type": "object",
        "properties": {
          "userCode": {
            "maxLength": 32,
            "type": "string"
          }
        }
      },
...
    }
  },
...
}